Improve reconnection logic
The only way to know if the client's connection was lost was either
through the consumer's shutdown and cancel callbacks or by calling publish
and receiving an exception due to the loss of connection. This refactors
the rabbit consumer and handler to conform to this logic.

Previously multiple threads could call enableRabbit to reconnect in a
multithreaded system (BungeeCord). Synchronized reconnections.